+++ .babelrc
@@ -1,6 +1,5 @@
 {
   "presets": ["@babel/preset-env", "@babel/preset-react"],
   "plugins": [
-    "styled-jsx/babel"
   ]
 }
