+++ package.json
@@ -22,16 +22,17 @@
   },
   "homepage": "https://github.com/elenamik/SmallReminders#readme",
   "dependencies": {
+    "@types/react": "^16.9.17",
     "axios": "^0.19.0",
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
+    "eslint-config-standard": "^14.1.0",
     "express": "^4.17.1",
     "mongodb": "^3.4.0",
     "mongoose": "^5.8.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
-    "styled-jsx": "^3.2.4",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10",
     "webpack-merge": "^4.2.2"
@@ -43,13 +44,17 @@
     "babel-loader": "^8.0.6",
     "babel-polyfill": "^6.26.0",
     "clean-webpack-plugin": "^3.0.0",
+    "css-loader": "^3.4.2",
     "firebase-tools": "^7.11.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^24.9.0",
+    "node-sass": "^4.13.0",
     "nodemon": "^2.0.2",
     "react-hot-loader": "^4.12.18",
+    "sass-loader": "^8.0.1",
     "semistandard": "^14.2.0",
+    "style-loader": "^1.1.2",
     "supertest": "^4.0.2",
     "webpack-dev-server": "^3.10.1"
   },
