+++ public/index.html
@@ -2,8 +2,7 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta.2/css/bootstrap.min.css" >
-    <title>Small Reminders</title>
+    <link href="https://fonts.googleapis.com/css?family=Libre+Baskerville&display=swap" rel="stylesheet">    <title>Small Reminders</title>
   </head>
   <body>
     <div id="app"></div>
@@ -25,5 +24,6 @@
       // Initialize Firebase
       firebase.initializeApp(firebaseConfig);
   </script>
+  <link href="https://fonts.googleapis.com/css?family=Gelasio&display=swap" rel="stylesheet">
   </body>
 </html>
