+++ src/client/components/Header/Header.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-
+import './Header.scss';
 function Header () {
   return (
-    <div id='app'>
+    <div id='header'>
       <h1>Small Reminders</h1>
     </div>
   );
