+++ src/client/components/Header/Header.scss
@@ -1,0 +1,8 @@
+@import '../../styles/colors.scss';
+#header {
+  color: $complement-4;
+  background: $complement-1;
+  min-height: 70px;
+  padding-left: 5%;
+  padding-top: 1.4em;
+}
