+++ src/client/components/Principle/index.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './Principle.scss';
+
+function Principle (props) {
+  return (
+    <div id='principle'>
+      {props.content}
+    </div>
+  );
+}
+
+export default Principle;
