+++ src/client/styles/colors.scss
@@ -1,0 +1,11 @@
+$primary-0: rgba( 72, 91,143,1);	// Main Primary color */
+$primary-1: rgb(179, 198, 250);
+$primary-2: rgba( 97,117,170,1);
+$primary-3: rgba( 48, 65,111,1);
+$primary-4: rgba( 26, 42, 88,1);
+
+$complement-0: rgba(212,174, 94,1);	// Main Complement color */
+$complement-1: rgba(255,219,143,1);
+$complement-2: rgba(253,214,131,1);
+$complement-3: rgba(164,131, 59,1);
+$complement-4: rgba(130, 97, 27,1);
