+++ src/client/views/Main/Main.scss
@@ -1,0 +1,7 @@
+#principles-list {
+  display: flex;
+  align-content: center;
+  justify-content: center;
+  flex-direction: column;
+  flex-flow: row wrap;
+}
