+++ src/client/views/Main/PrinciplesList.scss
@@ -1,0 +1,6 @@
+#principles-list {
+  display: flex;
+  align-content: center;
+  justify-content: center;
+  flex-direction: row;
+}
